Add IntDictionaryCompilerSmallData (#142)
* - added IntDictionaryCompilerSmallData

* - use uint64

* - rollback version

* - use IntInnerWeightsValueStoreBase::value_t = uint64_t;

* - fix style

+++ keyvi/include/keyvi/dictionary/fsa/internal/int_value_store.h
@@ -41,7 +41,7 @@ namespace internal {
 
 class IntValueStoreBase {
  public:
-  typedef uint32_t value_t;
+  typedef uint64_t value_t;
   static const bool inner_weight = false;
 
   IntValueStoreBase() {}
